\name{is.ra}
\alias{is.ra}
\title{Am I running Ra?}
\description{

After loading the \code{jit} library,
\code{is.ra} is \code{TRUE} if running Ra, else \code{FALSE}.

Note also that \link{version} will contain "\code{Ra}" when running Ra.
}
\usage{
  is.ra
}
\seealso{
    \code{\link{jit}}\cr
    \code{\link{nojit}}
}
\examples{
is.ra
}
\keyword{programming}
